Add files via upload
Implementation of a Deep Q-Network 

The local q-network (line 28) plays a major role in the DQN as it trains itself using gradients optimization algorithms (line 31) that will compute the loss (line 57) to further minimising it hence optimising the agent learning. Furthermore, the Bellman equation was introduced (line 83) taking into account the next state actions, whilst computing the expected Q-value. This will reduce the loss for the next transitions, hence allowing the agent to make the most efficient decisions. A mini-batch was implemented to sample a certain number of data (in our case a 100) (line 102) and stores them into a replay buffer. Once the replay buffer gets enough data, the q-network then trains the mini-batch (line 203). As mentioned in Question 1.1, this method was the most stable and efficient to compute the loss within the network. Furthermore, creating a second q-network, the ’target network’ (line 30). This one consists in an identical replication of the local q-network itself, as well as utilising its weights to be updated. The weights of the network predicts the effect specific inputs will have on the outputs. The Double DQN is therefore implemented to calculate the loss, where the local q-network trains the actual Q-values at a current state (line 77) and the target q-network trains the predicted Q-values at successor states (line 81). Therefore, both enable the loss to constantly decrease throughout predictions made. The greedy policy chooses the action with the highest Q-values (line 213), and tells the agent what decision to make. As the environment world we experimented in is a perfect representation of the real world, we first computed discrete actions within the four possible actions (line 173) and converted them into continuous actions with specific combinations of movements for the agent to execute. Furthermore, a decaying epsilon was set (line 140) for the q-network to have the best of both world, both exploration and exploitation of the agent.